requirements.txt
django
django-ninja
rav
pydantic[email]
django-ninja-jwt[crypto]
django-cors-headers